Capture replaces the vitals Record page; custom-vital management moves to configuration
- /care/vitals now renders the capture experience embedded in the
  admin shell (hamburger nav, sidebar patient selection) via a shared
  CaptureVitalsPanel; the standalone phone shell stays at /capture
- AdminV2Vitals is history-only (record form and inline definition
  creation removed; RecordVitalsForm remains for the dashboard quick-add)
- New CustomVitalsCard on the patient configuration page (add/remove
  definitions); VitalRangesCard refreshes to pick up new vitals
- Draft readings for since-deleted custom vitals are dropped at save
  instead of failing the batch
